requirements.txt
Flask>=2.0
python-chess>=1.999
pytest>=7.0
